Enables CGO free build
Enables building using `modernc.org/sqlite` without using CGO.
The default is to build with CGO enabled,
as performance is better using CGO.
