+++ importer_buffer_test.go
@@ -81,7 +81,7 @@ func TestBufferImporter_Import(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			db, err := Connect("sqlite3", "")
+			db, err := Connect(DefaultDriver, "")
 			if err != nil {
 				t.Fatal(err)
 			}
