+++ importer_slice_test.go
@@ -90,7 +90,7 @@ func TestSliceImporter_Import(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			db, err := Connect("sqlite3", "")
+			db, err := Connect(DefaultDriver, "")
 			if err != nil {
 				t.Fatal(err)
 			}
